fast_api/app/pyproject.toml
[tool.poetry]
name = "fast-api"
version = "0.1.0"
description = "API com FastAPI"
authors = ["DiogoBotton <diogo.luo@gmail.com>"]
#readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.2,<0.35.0"
sqlalchemy = ">=2.0.41,<3.0.0"
alembic = ">=1.15.2,<2.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
pydantic-settings = ">=2.9.1,<3.0.0"
validate-docbr = ">=1.10.0,<2.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"